Wrap bug fix and easier wrap config
Set up an event in player to configure the master room, allowing designers to run this event easily in their levels' level blueprints in BeginPlay.
Fixed an issue with level wrapping that caused infinite loop bug.
Aware of game-breaking issue with level wrapping that can cause player to be pushed out of the level bounds and have a plan to fix the problem.